Separated emails into agenda
